Initialize project with UrlParser and UrlInstance models
Create a ActiveRecord model UrlInstance that will store each instance of
a URL in the  database. This currently is only the shorthand and
longhand of the URL, but  will likely be extended with a user ID and
count.

The UrlParser is a helper model that handles any logic to create
shorthand versions if considered valid