Add is_valid_url? to UrlParser to ensure valid URLs
To ensure we are working with URLs, we use Ruby's internal URI::regexp
method to check whether we consider the URL to be valid. We currently do
not check if the URL actually exists, only that its layout passes our
criteria.

We check is_valid_url? against email address, random string, and an
instance with trailing whitespaces. We expect it to fail in all
instances.