Update README with first instructions and assumptions
Set up the README with all current knowledge on the problem space. Laid
out are all the assumptions made to keep track of, and the possible user
stories that indicate any work left to consider the project "finished".

The user stories are only an indication of current understanding, and
can be revisited and adapted in the future.