Create UrlInstances Controller with #create to create new UrlInstance objects
This commit purely focuses on a happy-path implementation of creating a 
UrlInstance through #create. It does not consider validity or 
error-handling. That will be addressed in a future commit.

I also realise we do not need to initialise the helper model UrlParser. 
For this reason the methods are made class methods.