Initialise Sessions and Home controller
initialise both the sessions and home controller to keep track if a user 
is logged in or out. The home controller is the root endpoint a user 
will be redirected to after login.

Most instances are fully set up by rails. I have added the routes to 
config/routes.rb to have more friendly endpoints in our routing