Ensure user is logged in before it uses the UrlInstance controller
Currently you can only access the URlInstance controller when you are a
logged in user. This means that you need to be logged in to both create
new shorthand URLs and make GET requests to these shorthands.

I am considering to exclude the GET /short/:shorthand endpoint from
this, so any anonymous user can use the shorthands.