Update /short/:shorthand endpoint to be allowed to be used anonymously
To ensure anyone can use the /short/:shorthand endpoint, and you can use 
your shortened URLs with any user, the :login_required helper method is 
not applied to the `#redirect_from_short` method.

To be clear in the tests that we are redirecting to the expected URL and 
not to home, we clarify this in the tests.