Add `count` column to UrlInstance, increased each time `/short/` endpoint is called
Each time someone makes a request to an existing shorthand endpoint, we 
want to increase the count to keep track how often each one is used.

Because we have decided anonymous users can use the shorthand endpoints, 
we need to make the relationship urlInstance->user optional to allow an 
anonymous user to increase the count.

There is room to consider alternative options here and maintain the 
user/url relationship.