Update rendering / redirecting of url_instances controller, update specs to reflect this
Previously there wasn't consistent behaviour through the app when an 
unsuccessful call was made;  sometimes the controller would render an 
unsuccessful body and sometimes it would redirect with a flash notice.

To keep the flow more consistent, I have decided to redirect in these 
scenarios with a flash notice. If I would have decided to return JSON 
blobs from the controllers, non-OK responses would be more accurate to 
use.